like and dislike product

+++ src/services/DeleteProductService.js
@@ -2,6 +2,16 @@ const { Product } = require('../models/Product');
 
 class DeleteProductService {
   async execute(userId, productId) {
+    const result = await Product.findOne({
+      where: {
+        id: productId,
+      },
+    });
+
+    if (!result) {
+      return new Error('This product does not exits');
+    }
+
     await Product.destroy({
       where: {
         user_id: userId,
